requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
antlr4-python3-runtime=4.8=pypi_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py310h7f8727e_0
asttokens=2.0.5=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
autoprop=4.0.2=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py310h06a4308_0
black=22.3.0=pypi_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
brotlipy=0.7.0=py310h7f8727e_1002
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.10.8=ha878542_0
certifi=2021.10.8=py310hff52083_2
cffi=1.15.0=py310hd667e15_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=pypi_0
colorama=0.4.4=pyh9f0ad1d_0
commonmark=0.9.1=pypi_0
cryptography=36.0.0=py310h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pypi_0
cython=0.29.20=pypi_0
dataclasses=0.8=pyhc8e2a94_3
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py310h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.4=py310h06a4308_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.4.4=h295c915_0
ffmpeg=4.3=hf484d3e_0
filelock=3.6.0=pypi_0
flake8=4.0.1=pypi_0
flake8-bugbear=22.4.25=pypi_0
flake8-comprehensions=3.8.0=pypi_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.33.3=pypi_0
freetype=2.11.0=h70c0345_0
fvcore=0.1.5.post20220506=pyhd8ed1ab_0
gdown=4.4.0=pypi_0
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h4ff587b_1
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
hydra-core=1.1.2=pypi_0
icu=58.2=he6710b0_3
idna=3.3=pyhd3eb1b0_0
imageio=2.19.0=pypi_0
iniconfig=1.1.1=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
iopath=0.1.9=pypi_0
ipykernel=6.9.1=py310h06a4308_0
ipython=8.3.0=py310h06a4308_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=py310h06a4308_1
jinja2=3.0.3=pyhd3eb1b0_0
jpeg=9e=h7f8727e_0
jsonschema=4.4.0=py310h06a4308_0
jupyter=1.0.0=py310h06a4308_7
jupyter_client=7.2.2=py310h06a4308_0
jupyter_console=6.4.3=pyhd3eb1b0_0
jupyter_core=4.10.0=py310h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.4.2=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libcst=0.4.2=pypi_0
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h74e7548_1
lz4-c=1.9.3=h295c915_1
markupsafe=2.0.1=py310h7f8727e_0
matplotlib=3.5.2=pypi_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mccabe=0.6.1=pypi_0
mistune=0.8.4=py310h7f8727e_1000
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
mkl_fft=1.3.1=py310hd6ae3a3_0
mkl_random=1.2.2=py310h00e6091_0
more-itertools=8.13.0=pypi_0
moreorless=0.4.0=pypi_0
mpmath=1.2.1=pypi_0
mypy-extensions=0.4.3=pypi_0
nbclient=0.5.13=py310h06a4308_0
nbconvert=6.4.4=py310h06a4308_0
nbformat=5.3.0=py310h06a4308_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.5=py310h06a4308_0
nettle=3.7.3=hbbd107a_1
networkx=2.8=pypi_0
notebook=6.4.11=py310h06a4308_0
numpy=1.21.5=py310hfa59a62_2
numpy-base=1.21.5=py310h9585f30_2
omegaconf=2.1.2=pypi_0
opencv-python-headless=4.5.5.64=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1n=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pathspec=0.9.0=pypi_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.0.1=py310h22f2fdc_0
pip=22.0.4=pypi_0
platformdirs=2.5.2=pypi_0
plotly=5.7.0=pypi_0
pluggy=1.0.0=pypi_0
portalocker=2.4.0=py310hff52083_0
prometheus_client=0.13.1=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
py=1.11.0=pypi_0
pycodestyle=2.8.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyflakes=2.4.0=pypi_0
pyglet=1.5.23=pypi_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.8=pypi_0
pyqt=5.9.2=py310h295c915_6
pyrsistent=0.18.0=py310h7f8727e_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.4=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.15.1=pyhd3eb1b0_0
python_abi=3.10=2_cp310
pytorch=1.11.0=py3.10_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytorch3d=0.6.2=py310_cu113_pyt1110
pywavelets=1.3.0=pypi_0
pyyaml=6.0=pypi_0
pyzmq=22.3.0=py310h295c915_2
qt=5.9.7=h5867ecd_1
qtconsole=5.3.0=pyhd3eb1b0_0
qtpy=2.0.1=pyhd3eb1b0_0
readline=8.1.2=h7f8727e_1
requests=2.27.1=pyhd3eb1b0_0
rich=12.4.1=pypi_0
rtree=1.0.0=pypi_0
scikit-image=0.19.2=pypi_0
scipy=1.8.0=pypi_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=61.2.0=py310h06a4308_0
shapely=1.8.2=pypi_0
signature-dispatch=1.0.0=pypi_0
sip=4.19.13=py310h295c915_0
six=1.16.0=pyhd3eb1b0_1
soupsieve=2.3.2.post1=pypi_0
sqlite=3.38.3=hc218d9a_0
stack_data=0.2.0=pyhd3eb1b0_0
stdlibs=2022.3.16=pypi_0
svg-path=6.0=pypi_0
sympy=1.10.1=pypi_0
tabulate=0.8.9=pyhd8ed1ab_0
tenacity=8.0.1=pypi_0
termcolor=1.1.0=py_2
terminado=0.13.1=py310h06a4308_0
testpath=0.5.0=pyhd3eb1b0_0
tifffile=2022.5.4=pypi_0
tk=8.6.11=h1ccaba5_0
toml=0.10.2=pypi_0
tomli=2.0.1=pypi_0
torchaudio=0.11.0=py310_cu113
torchvision=0.12.0=py310_cu113
tornado=6.1=py310h7f8727e_0
tqdm=4.64.0=pyhd8ed1ab_0
trailrunner=1.1.3=pypi_0
traitlets=5.1.1=pyhd3eb1b0_0
triangle=20220202=pypi_0
trimesh=3.12.0=pypi_0
typeguard=2.13.3=pypi_0
typing-extensions=4.1.1=hd3eb1b0_0
typing-inspect=0.7.1=pypi_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.9=py310h06a4308_0
usort=1.0.2=pypi_0
vecrec=0.3.1=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py310h06a4308_1
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py310h06a4308_0
xatlas=0.0.6=pypi_0
xxhash=3.0.0=pypi_0
xz=5.2.5=h7f8727e_1
yacs=0.1.8=pyhd8ed1ab_0
yaml=0.2.5=h516909a_0
zeromq=4.3.4=h2531618_0
zlib=1.2.12=h7f8727e_2
zstd=1.4.9=haebb681_0
